+++ src/OpenTelemetry/Logs/LoggerProviderSdk.cs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Diagnostics;
-#if NETSTANDARD2_1_OR_GREATER || NET6_0_OR_GREATER
+#if NETSTANDARD2_1_OR_GREATER || NET
 using System.Diagnostics.CodeAnalysis;
 #endif
 using System.Text;
@@ -201,7 +201,7 @@ internal sealed class LoggerProviderSdk : LoggerProvider
 #endif
         override bool TryCreateLogger(
         string? name,
-#if NETSTANDARD2_1_OR_GREATER || NET6_0_OR_GREATER
+#if NETSTANDARD2_1_OR_GREATER || NET
         [NotNullWhen(true)]
 #endif
         out Logger? logger)